Add a :mashify option to turn off the Mashify middleware
This option makes it possible to get raw responses without directly
modifying the middleware.

The Mashify middleware is not inserted if the :mashify option == false.
The default is nil to retain back compatibility.

The Restforce::Concerns::Connection#mashify? method is unmodified. It still
looks at the middleware stack to determine if it includes Mashify, so there
should be no change in behavior for existing uses that modify the
stack to remove Mashify.