set SF api version from env var
All other options have a related ENV var from the system that's hosting this gem.  api_version defaults to 26.0 but there's no way to set this from an ENV var upon deployment.  This change will allow the variable to be set or it will default to v26.0 if none was specified.