Relax `json` dependency to support Ruby 2.4.0-dev
Ruby 2.4.0+ requires a newer version of the `json` gem which is incompatible with pre-2.0.0 versions (i.e. 1.9.3, which the gem supports). This changes the gemspec to be able to handle both cases. Tests pass with both versions of the dependency to there's no problem.

This PR also allows `rbx` to fail builds - apparently it's broken in Travis at the moment (see <https://github.com/vmg/redcarpet/issues/580>).

Fixes #260.