Track replayIds
This allows the replay option passed to remember the last replayId and
to allow each channel to have it's own replay_id

The replay object supports hash-like access and can be used to provide
static values per channel using a Hash. Or if dynamic behaviour is
needed the caller can provide a custom object that responds to [] and
[]=

The implementation is *not* thread-safe and it is probably a terrible
idea to use a hash here.